requirements.txt
python-telegram-bot 
boto3 
openai 
requests 
matplotlib 
pytz
thingspeak
Flask
gunicorn